backend/requirements.txt
alembic==1.14.1
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
fastapi==0.115.8
uvicorn==0.34.0
passlib==1.7.4
pyjwt==2.10.1
pytest==8.3.4
httpx==0.28.1
bcrypt==4.2.1
pydantic[email]==2.10.6
